src/app/photo.service.ts
import { Injectable } from '@angular/core';
import { HttpClient, HttpParams } from '@angular/common/http';
import { Observable } from 'rxjs';

@Injectable({
  providedIn: 'root'
})
export class PhotoService {
  apiUrl = 'https://jsonplaceholder.typicode.com/photos';

  constructor(private http: HttpClient) {}

  getPhotos(page: number, limit: number): Observable<any[]> {
    let params = new HttpParams()
      .set('_page', page.toString())
      .set('_limit', limit.toString())

    return this.http.get<any[]>(this.apiUrl, { params });
  }
}
